requirements.txt
-c requirements-constraints.txt

numpy~=1.21
monai[nibabel, einops]~=1.3.1
scipy~=1.10
fastremap~=1.13
opencv-python~=4.8
pillow~=10.0
tqdm~=4.64
humanize~=4.9.0


nibabel~=5.2.1
pytorch-ignite~=0.5
pynvml
psutil
comet-ml
